LayoutTests/platform/ios-sim-deprecated/transforms/3d/hit-testing/perspective-clipped-expected.txt

CONSOLE MESSAGE: line 15: TypeError: null is not an object (evaluating 'hit.id')
Element at 35, 100 has id "topLayer": PASS
